+++ src/main/java/duke/TaskList.java
@@ -1,0 +1,5 @@
+package duke;
+
+public class TaskList {
+
+}
